Removed the ability to delete objects, where necessary and possible, that have associated records in the database in order to maintain data integrity. Added navigation between login, blog and personal account.

+++ app/Http/Controllers/Admin/Category/DeleteController.php
@@ -9,6 +9,9 @@ class DeleteController extends Controller
 {
     public function __invoke(Category $category)
     {
+        if ($category->posts()->exists()) {
+            throw new \Exception('Cannot delete, related records exist.');
+        }
         $category->delete();
         return redirect()->route('admin.category.index', compact('category'));
     }
